+++ lib/src/models/excluded_payment_methods/excluded_payment_methods.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: public_member_api_docs
 import 'package:json_annotation/json_annotation.dart';
 
 part 'excluded_payment_methods.g.dart';