+++ lib/src/models/excluded_payment_types/excluded_payment_types.dart
@@ -1,17 +1,17 @@
+// ignore_for_file: public_member_api_docs
 import 'package:json_annotation/json_annotation.dart';
 
 part 'excluded_payment_types.g.dart';
 
 @JsonSerializable()
 class ExcludedPaymentTypes {
+  factory ExcludedPaymentTypes.fromJson(Map<String, dynamic> json) =>
+      _$ExcludedPaymentTypesFromJson(json);
   ExcludedPaymentTypes({
     this.id,
   });
 
   String? id;
-
-  factory ExcludedPaymentTypes.fromJson(Map<String, dynamic> json) =>
-      _$ExcludedPaymentTypesFromJson(json);
 
   Map<String, dynamic> toJson() => _$ExcludedPaymentTypesToJson(this);
 }