+++ lib/src/mp_exception.dart
@@ -1,7 +1,0 @@
-class MpException implements Exception {
-  MpException({
-    required this.message,
-  });
-
-  final String message;
-}